Add design changes
