--- src/app/api/auth/check/route.js
@@ -1,18 +1,0 @@
-import { NextResponse } from "next/server";
-import { cookies } from "next/headers";
-
-export async function GET() {
-  const cookieStore = await cookies();
-  const userCookie = cookieStore.get("user");
-
-  if (!userCookie) {
-    return NextResponse.json({ authenticated: false });
-  }
-
-  try {
-    const user = JSON.parse(userCookie.value);
-    return NextResponse.json({ authenticated: true, user });
-  } catch (error) {
-    return NextResponse.json({ authenticated: false });
-  }
-}
