--- src/app/api/auth/logout/route.js
@@ -1,8 +1,0 @@
-import { NextResponse } from "next/server";
-import { cookies } from "next/headers";
-
-export async function POST() {
-  const cookieStore = await cookies();
-  cookieStore.delete("user");
-  return NextResponse.json({ success: true });
-}
