--- src/app/api/auth/register/route.js
@@ -1,64 +1,0 @@
-import { NextResponse } from "next/server";
-import { cookies } from "next/headers";
-
-export async function POST(request) {
-  try {
-    const { firstName, lastName, email, password } = await request.json();
-
-    // In a real application, you would:
-    // 1. Validate the input data
-    // 2. Check if the email is already registered
-    // 3. Hash the password
-    // 4. Store the user in the database
-    // For now, we'll use a simple mock registration
-
-    // Mock validation
-    if (!firstName || !lastName || !email || !password) {
-      return NextResponse.json(
-        { error: "All fields are required" },
-        { status: 400 }
-      );
-    }
-
-    // Mock email format validation
-    if (!email.includes("@")) {
-      return NextResponse.json(
-        { error: "Invalid email format" },
-        { status: 400 }
-      );
-    }
-
-    // Mock password validation
-    if (password.length < 6) {
-      return NextResponse.json(
-        { error: "Password must be at least 6 characters" },
-        { status: 400 }
-      );
-    }
-
-    // Create a mock user
-    const user = {
-      id: "2", // In a real app, this would be generated by the database
-      email,
-      firstName,
-      lastName,
-    };
-
-    // Set the user cookie
-    const cookieStore = cookies();
-    cookieStore.set("user", JSON.stringify(user), {
-      path: "/",
-      maxAge: 60 * 60 * 24 * 7, // 1 week
-      httpOnly: true,
-      secure: process.env.NODE_ENV === "production",
-    });
-
-    return NextResponse.json({ user });
-  } catch (error) {
-    console.error("Registration error:", error);
-    return NextResponse.json(
-      { error: "Internal server error" },
-      { status: 500 }
-    );
-  }
-}
