+++ src/app/login/page.js
@@ -1,269 +1,5 @@
-"use client";
-
-import { useState } from "react";
-import Link from "next/link";
-import Image from "next/image";
-import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
-import { Tabs, TabsList, TabsTrigger, TabsContent } from "@/components/ui/tabs";
-import FloatingCartButton from "@/components/FloatingCartButton";
+import Login from "@/components/Login";
 
 export default function LoginPage() {
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const [firstName, setFirstName] = useState("");
-  const [lastName, setLastName] = useState("");
-  const [confirmPassword, setConfirmPassword] = useState("");
-  const [agreeToTerms, setAgreeToTerms] = useState(false);
-
-  const handleLogin = async (e) => {
-    e.preventDefault();
-    try {
-      const response = await fetch("/api/auth/login", {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify({ email, password }),
-      });
-
-      if (response.ok) {
-        const data = await response.json();
-        // Set user cookie
-        document.cookie = `user=${JSON.stringify(data.user)}; path=/`;
-        window.location.href = "/";
-      } else {
-        alert("Login failed. Please check your credentials.");
-      }
-    } catch (error) {
-      console.error("Login error:", error);
-      alert("An error occurred during login.");
-    }
-  };
-
-  const handleRegister = async (e) => {
-    e.preventDefault();
-    if (password !== confirmPassword) {
-      alert("Passwords do not match!");
-      return;
-    }
-
-    try {
-      const response = await fetch("/api/auth/register", {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-        },
-        body: JSON.stringify({
-          firstName,
-          lastName,
-          email,
-          password,
-        }),
-      });
-
-      if (response.ok) {
-        const data = await response.json();
-        // Set user cookie
-        document.cookie = `user=${JSON.stringify(data.user)}; path=/`;
-        window.location.href = "/";
-      } else {
-        alert("Registration failed. Please try again.");
-      }
-    } catch (error) {
-      console.error("Registration error:", error);
-      alert("An error occurred during registration.");
-    }
-  };
-
-  return (
-    <div className="min-h-screen bg-background">
-      {/* Login Content */}
-      <div className="container mx-auto px-4 py-12 md:py-16">
-        <div className="max-w-md mx-auto">
-          <h1 className="text-center text-2xl text-neutral-dark font-normal mb-8">Account</h1>
-
-          <Tabs defaultValue="login" className="w-full">
-            <TabsList className="grid w-full grid-cols-2 mb-8 bg-secondary">
-              <TabsTrigger
-                value="login"
-                className="text-sm data-[state=active]:bg-primary data-[state=active]:text-white"
-              >
-                Login
-              </TabsTrigger>
-              <TabsTrigger
-                value="register"
-                className="text-sm data-[state=active]:bg-primary data-[state=active]:text-white"
-              >
-                Create Account
-              </TabsTrigger>
-            </TabsList>
-
-            <TabsContent value="login">
-              <form onSubmit={handleLogin} className="space-y-6">
-                <div className="space-y-2">
-                  <Label htmlFor="email" className="text-neutral-dark">
-                    Email
-                  </Label>
-                  <Input
-                    id="email"
-                    type="email"
-                    placeholder="your@email.com"
-                    value={email}
-                    onChange={(e) => setEmail(e.target.value)}
-                    required
-                    className="bg-white border-neutral-light text-neutral-dark"
-                  />
-                </div>
-                <div className="space-y-2">
-                  <div className="flex items-center justify-between">
-                    <Label htmlFor="password" className="text-neutral-dark">
-                      Password
-                    </Label>
-                    <Link href="/forgot-password" className="text-xs text-neutral-medium hover:text-primary">
-                      Forgot password?
-                    </Link>
-                  </div>
-                  <Input
-                    id="password"
-                    type="password"
-                    value={password}
-                    onChange={(e) => setPassword(e.target.value)}
-                    required
-                    className="bg-white border-neutral-light text-neutral-dark"
-                  />
-                </div>
-                <Button type="submit" className="w-full bg-primary hover:bg-accent text-white rounded-none">
-                  Sign In
-                </Button>
-
-                <div className="relative my-6">
-                  <div className="absolute inset-0 flex items-center">
-                    <div className="w-full border-t border-neutral-light"></div>
-                  </div>
-                  <div className="relative flex justify-center text-xs uppercase">
-                    <span className="bg-background px-2 text-neutral-medium">Or continue with</span>
-                  </div>
-                </div>
-
-                <div className="grid grid-cols-2 gap-4">
-                  <Button
-                    type="button"
-                    variant="outline"
-                    className="rounded-none border-neutral-light text-neutral-dark hover:bg-secondary"
-                  >
-                    <Image src="/images/google-icon.png" alt="Google" width={18} height={18} className="mr-2" />
-                    Google
-                  </Button>
-                  <Button
-                    type="button"
-                    variant="outline"
-                    className="rounded-none border-neutral-light text-neutral-dark hover:bg-secondary"
-                  >
-                    <Image src="/images/facebook-icon.png" alt="Facebook" width={18} height={18} className="mr-2" />
-                    Facebook
-                  </Button>
-                </div>
-              </form>
-            </TabsContent>
-
-            <TabsContent value="register">
-              <form onSubmit={handleRegister} className="space-y-6">
-                <div className="grid grid-cols-2 gap-4">
-                  <div className="space-y-2">
-                    <Label htmlFor="first-name" className="text-neutral-dark">
-                      First Name
-                    </Label>
-                    <Input
-                      id="first-name"
-                      value={firstName}
-                      onChange={(e) => setFirstName(e.target.value)}
-                      required
-                      className="bg-white border-neutral-light text-neutral-dark"
-                    />
-                  </div>
-                  <div className="space-y-2">
-                    <Label htmlFor="last-name" className="text-neutral-dark">
-                      Last Name
-                    </Label>
-                    <Input
-                      id="last-name"
-                      value={lastName}
-                      onChange={(e) => setLastName(e.target.value)}
-                      required
-                      className="bg-white border-neutral-light text-neutral-dark"
-                    />
-                  </div>
-                </div>
-                <div className="space-y-2">
-                  <Label htmlFor="register-email" className="text-neutral-dark">
-                    Email
-                  </Label>
-                  <Input
-                    id="register-email"
-                    type="email"
-                    placeholder="your@email.com"
-                    value={email}
-                    onChange={(e) => setEmail(e.target.value)}
-                    required
-                    className="bg-white border-neutral-light text-neutral-dark"
-                  />
-                </div>
-                <div className="space-y-2">
-                  <Label htmlFor="register-password" className="text-neutral-dark">
-                    Password
-                  </Label>
-                  <Input
-                    id="register-password"
-                    type="password"
-                    value={password}
-                    onChange={(e) => setPassword(e.target.value)}
-                    required
-                    className="bg-white border-neutral-light text-neutral-dark"
-                  />
-                </div>
-                <div className="space-y-2">
-                  <Label htmlFor="confirm-password" className="text-neutral-dark">
-                    Confirm Password
-                  </Label>
-                  <Input
-                    id="confirm-password"
-                    type="password"
-                    value={confirmPassword}
-                    onChange={(e) => setConfirmPassword(e.target.value)}
-                    required
-                    className="bg-white border-neutral-light text-neutral-dark"
-                  />
-                </div>
-                <div className="flex items-center space-x-2">
-                  <input
-                    type="checkbox"
-                    id="terms"
-                    className="rounded border-neutral-light bg-white accent-primary"
-                    checked={agreeToTerms}
-                    onChange={(e) => setAgreeToTerms(e.target.checked)}
-                    required
-                  />
-                  <Label htmlFor="terms" className="text-xs text-neutral-medium">
-                    I agree to the{" "}
-                    <Link href="/terms" className="underline text-primary">
-                      Terms of Service
-                    </Link>{" "}
-                    and{" "}
-                    <Link href="/privacy" className="underline text-primary">
-                      Privacy Policy
-                    </Link>
-                  </Label>
-                </div>
-                <Button type="submit" className="w-full bg-primary hover:bg-accent text-white rounded-none">
-                  Create Account
-                </Button>
-              </form>
-            </TabsContent>
-          </Tabs>
-        </div>
-      </div>
-    </div>
-  );
+  return <Login />;
 }
