+++ src/components/Login.js
@@ -3,20 +3,27 @@
 import { useState } from "react";
 import { useRouter } from "next/navigation";
 import Link from "next/link";
+import Image from "next/image";
+import { Button } from "@/components/ui/button";
+import { Input } from "@/components/ui/input";
+import { Label } from "@/components/ui/label";
+import { Tabs, TabsList, TabsTrigger, TabsContent } from "@/components/ui/tabs";
 import { useAuth } from "@/app/context/AuthContext";
 
-const backgroundImageUrl = "/ifr.jpg";
-
-const Login = () => {
+export default function Login() {
   const [formData, setFormData] = useState({
     email: "",
     password: "",
   });
-
-  const { login } = useAuth();
+  const [firstName, setFirstName] = useState("");
+  const [lastName, setLastName] = useState("");
+  const [confirmPassword, setConfirmPassword] = useState("");
+  const [agreeToTerms, setAgreeToTerms] = useState(false);
   const [showPassword, setShowPassword] = useState(false);
   const [error, setError] = useState("");
-  const [loading, setLoading] = useState(false); // Added loading state
+  const [loading, setLoading] = useState(false);
+  
+  const { login } = useAuth();
   const router = useRouter();
 
   const handleChange = (event) => {
@@ -30,13 +37,13 @@ const Login = () => {
   const handleSubmit = async (event) => {
     event.preventDefault();
     setError("");
+    setLoading(true);
 
     if (!formData.email || !formData.password) {
       setError("Email and password are required");
+      setLoading(false);
       return;
     }
-
-    setLoading(true); // Disable button & show loading
 
     try {
       const response = await fetch("/api/login", {
@@ -54,16 +61,14 @@ const Login = () => {
       if (!response.ok) {
         const data = await response.json();
         setError(data.error || "Invalid email or password");
-        setLoading(false); // Re-enable button on error
+        setLoading(false);
         return;
       }
 
       const data = await response.json();
-
       const userResponse = await fetch("/api/auth/user", {
         credentials: "include",
       });
-
       const userData = await userResponse.json();
 
       if (userData.isAuthenticated) {
@@ -76,93 +81,250 @@ const Login = () => {
     } catch (error) {
       setError("An error occurred. Please try again.");
       console.error(error);
-      setLoading(false); // Re-enable button on error
+      setLoading(false);
+    }
+  };
+
+  const handleRegister = async (e) => {
+    e.preventDefault();
+    if (formData.password !== confirmPassword) {
+      alert("Passwords do not match!");
+      return;
+    }
+
+    try {
+      const response = await fetch("/api/auth/register", {
+        method: "POST",
+        headers: {
+          "Content-Type": "application/json",
+        },
+        body: JSON.stringify({
+          firstName,
+          lastName,
+          email: formData.email,
+          password: formData.password,
+        }),
+      });
+
+      if (response.ok) {
+        const data = await response.json();
+        login(data.user);
+        router.push("/");
+      } else {
+        alert("Registration failed. Please try again.");
+      }
+    } catch (error) {
+      console.error("Registration error:", error);
+      alert("An error occurred during registration.");
     }
   };
 
   return (
-    <section
-      className="flex items-center justify-center min-h-screen bg-cover bg-center px-4"
-      style={{ backgroundImage: `url(${backgroundImageUrl})` }}
-    >
-      <div className="relative w-full max-w-4xl min-h-[600px] bg-white shadow-lg rounded-xl flex flex-col md:flex-row overflow-hidden">
-        {/* Left Panel: Sign In Form */}
-        <div className="w-full md:w-1/2 flex flex-col justify-center items-center p-8 bg-white bg-opacity-90 rounded-l-xl">
-          <h2 className="text-3xl font-bold text-gray-900 mb-8 mt-4">Sign in</h2>
+    <div className="min-h-screen bg-background">
+      <div className="container mx-auto px-4 py-12 md:py-16">
+        <div className="max-w-md mx-auto">
+          <h1 className="text-center text-2xl text-neutral-dark font-normal mb-8">Account</h1>
 
-          <form onSubmit={handleSubmit} className="w-full">
-            {/* Email Input */}
-            <input
-              type="email"
-              name="email"
-              placeholder="name@company.com"
-              value={formData.email}
-              onChange={handleChange}
-              className="w-full px-4 py-3 mb-4 bg-gray-100 border border-gray-700 text-gray-900 rounded-lg focus:ring-2 focus:ring-[#605137] placeholder-gray-400 transition-all"
-              required
-              disabled={loading} // Disable input when loading
-            />
-
-            {/* Password Input with Show/Hide Button */}
-            <div className="relative w-full">
-              <input
-                type={showPassword ? "text" : "password"}
-                name="password"
-                placeholder="Enter your password"
-                value={formData.password}
-                onChange={handleChange}
-                className="w-full px-4 py-3 bg-gray-100 border border-gray-700 text-gray-900 rounded-lg focus:ring-2 focus:ring-[#605137] placeholder-gray-400 transition-all pr-16"
-                required
-                disabled={loading} // Disable input when loading
-              />
-              <button
-                type="button"
-                onClick={() => setShowPassword((prev) => !prev)}
-                className="absolute inset-y-0 right-4 flex items-center text-sm text-[#605137] font-semibold hover:underline focus:outline-none"
-                disabled={loading} // Disable toggle when loading
+          <Tabs defaultValue="login" className="w-full">
+            <TabsList className="grid w-full grid-cols-2 mb-8 bg-secondary">
+              <TabsTrigger
+                value="login"
+                className="text-sm data-[state=active]:bg-primary data-[state=active]:text-white"
               >
-                {showPassword ? "Hide" : "Show"}
-              </button>
-            </div>
+                Login
+              </TabsTrigger>
+              <TabsTrigger
+                value="register"
+                className="text-sm data-[state=active]:bg-primary data-[state=active]:text-white"
+              >
+                Create Account
+              </TabsTrigger>
+            </TabsList>
 
-            {/* Display Error Message */}
-            {error && <p className="text-red-500 text-sm mt-2">{error}</p>}
+            <TabsContent value="login">
+              <form onSubmit={handleSubmit} className="space-y-6">
+                <div className="space-y-2">
+                  <Label htmlFor="email" className="text-neutral-dark">
+                    Email
+                  </Label>
+                  <Input
+                    id="email"
+                    type="email"
+                    name="email"
+                    placeholder="name@company.com"
+                    value={formData.email}
+                    onChange={handleChange}
+                    required
+                    disabled={loading}
+                    className="bg-white border-neutral-light text-neutral-dark"
+                  />
+                </div>
+                <div className="space-y-2">
+                  <div className="flex items-center justify-between">
+                    <Label htmlFor="password" className="text-neutral-dark">
+                      Password
+                    </Label>
+                    <Link href="/forgot-password" className="text-xs text-neutral-medium hover:text-primary">
+                      Forgot password?
+                    </Link>
+                  </div>
+                  <div className="relative">
+                    <Input
+                      id="password"
+                      type={showPassword ? "text" : "password"}
+                      name="password"
+                      placeholder="Enter your password"
+                      value={formData.password}
+                      onChange={handleChange}
+                      required
+                      disabled={loading}
+                      className="bg-white border-neutral-light text-neutral-dark pr-16"
+                    />
+                    <button
+                      type="button"
+                      onClick={() => setShowPassword((prev) => !prev)}
+                      className="absolute inset-y-0 right-4 flex items-center text-sm text-primary font-semibold hover:underline focus:outline-none"
+                      disabled={loading}
+                    >
+                      {showPassword ? "Hide" : "Show"}
+                    </button>
+                  </div>
+                </div>
+                {error && <p className="text-red-500 text-sm mt-2">{error}</p>}
+                <Button 
+                  type="submit" 
+                  className="w-full bg-primary hover:bg-accent text-white rounded-none"
+                  disabled={loading}
+                >
+                  {loading ? "Signing In..." : "Sign In"}
+                </Button>
 
-            {/* Forgot Password */}
-            <p className="text-sm text-gray-600 mb-4 mt-2">
-              Forgot your password?{" "}
-              <Link href="/forgot-password" className="text-[#605137] font-semibold hover:text-[#30291C] transition">
-                Click here
-              </Link>
-            </p>
+                <div className="relative my-6">
+                  <div className="absolute inset-0 flex items-center">
+                    <div className="w-full border-t border-neutral-light"></div>
+                  </div>
+                  <div className="relative flex justify-center text-xs uppercase">
+                    <span className="bg-background px-2 text-neutral-medium">Or continue with</span>
+                  </div>
+                </div>
 
-            {/* Submit Button with Loading State */}
-            <button
-              type="submit"
-              disabled={loading}
-              className={`w-full py-2 bg-[#30291C] text-white font-bold rounded-full mt-4 ${
-                loading ? "opacity-50 cursor-not-allowed" : ""
-              }`}
-            >
-              {loading ? "Signing In..." : "Sign In"}
-            </button>
-          </form>
-        </div>
+                <div className="grid grid-cols-2 gap-4">
+                  <Button
+                    type="button"
+                    variant="outline"
+                    className="rounded-none border-neutral-light text-neutral-dark hover:bg-secondary"
+                  >
+                    <Image src="/images/google-icon.png" alt="Google" width={18} height={18} className="mr-2" />
+                    Google
+                  </Button>
+                  <Button
+                    type="button"
+                    variant="outline"
+                    className="rounded-none border-neutral-light text-neutral-dark hover:bg-secondary"
+                  >
+                    <Image src="/images/facebook-icon.png" alt="Facebook" width={18} height={18} className="mr-2" />
+                    Facebook
+                  </Button>
+                </div>
+              </form>
+            </TabsContent>
 
-        {/* Right Panel: Sign Up Prompt */}
-        <div className="w-full md:w-1/2 flex flex-col justify-center items-center p-16 bg-[#605137] text-white rounded-r-xl">
-          <h2 className="text-3xl font-bold">New Here?</h2>
-          <p className="text-center mt-2">Create an account to start shopping with us.</p>
-
-          <Link href="/register">
-            <button className="mt-4 px-6 py-2 border border-white rounded-full hover:bg-white hover:text-[#605137] transition">
-              Create Account
-            </button>
-          </Link>
+            <TabsContent value="register">
+              <form onSubmit={handleRegister} className="space-y-6">
+                <div className="grid grid-cols-2 gap-4">
+                  <div className="space-y-2">
+                    <Label htmlFor="first-name" className="text-neutral-dark">
+                      First Name
+                    </Label>
+                    <Input
+                      id="first-name"
+                      value={firstName}
+                      onChange={(e) => setFirstName(e.target.value)}
+                      required
+                      className="bg-white border-neutral-light text-neutral-dark"
+                    />
+                  </div>
+                  <div className="space-y-2">
+                    <Label htmlFor="last-name" className="text-neutral-dark">
+                      Last Name
+                    </Label>
+                    <Input
+                      id="last-name"
+                      value={lastName}
+                      onChange={(e) => setLastName(e.target.value)}
+                      required
+                      className="bg-white border-neutral-light text-neutral-dark"
+                    />
+                  </div>
+                </div>
+                <div className="space-y-2">
+                  <Label htmlFor="register-email" className="text-neutral-dark">
+                    Email
+                  </Label>
+                  <Input
+                    id="register-email"
+                    type="email"
+                    placeholder="your@email.com"
+                    value={formData.email}
+                    onChange={handleChange}
+                    required
+                    className="bg-white border-neutral-light text-neutral-dark"
+                  />
+                </div>
+                <div className="space-y-2">
+                  <Label htmlFor="register-password" className="text-neutral-dark">
+                    Password
+                  </Label>
+                  <Input
+                    id="register-password"
+                    type="password"
+                    value={formData.password}
+                    onChange={handleChange}
+                    required
+                    className="bg-white border-neutral-light text-neutral-dark"
+                  />
+                </div>
+                <div className="space-y-2">
+                  <Label htmlFor="confirm-password" className="text-neutral-dark">
+                    Confirm Password
+                  </Label>
+                  <Input
+                    id="confirm-password"
+                    type="password"
+                    value={confirmPassword}
+                    onChange={(e) => setConfirmPassword(e.target.value)}
+                    required
+                    className="bg-white border-neutral-light text-neutral-dark"
+                  />
+                </div>
+                <div className="flex items-center space-x-2">
+                  <input
+                    type="checkbox"
+                    id="terms"
+                    className="rounded border-neutral-light bg-white accent-primary"
+                    checked={agreeToTerms}
+                    onChange={(e) => setAgreeToTerms(e.target.checked)}
+                    required
+                  />
+                  <Label htmlFor="terms" className="text-xs text-neutral-medium">
+                    I agree to the{" "}
+                    <Link href="/terms" className="underline text-primary">
+                      Terms of Service
+                    </Link>{" "}
+                    and{" "}
+                    <Link href="/privacy" className="underline text-primary">
+                      Privacy Policy
+                    </Link>
+                  </Label>
+                </div>
+                <Button type="submit" className="w-full bg-primary hover:bg-accent text-white rounded-none">
+                  Create Account
+                </Button>
+              </form>
+            </TabsContent>
+          </Tabs>
         </div>
       </div>
-    </section>
+    </div>
   );
-};
-
-export default Login;
+}
